requirements.txt
flask
requests
python-dotenv
flask-sqlalchemy

